+++ plugins/source/aws/resources/services/s3/access_points_fetch.go
@@ -1,4 +1,4 @@
-package s3control
+package s3
 
 import (
 	"context"
