+++ plugins/source/aws/resources/services/s3/access_points_mock_test.go
@@ -1,4 +1,4 @@
-package s3control
+package s3
 
 import (
 	"testing"
